Finished string refactor (#3)
